Added the Player input.
It handles input from keyboard and joystick in order
to do something on the player. These functions are almost the same
as the functions provided by Game, that has the responsibility.

Game is a gob object. It does too many things.
Game must do only on thing: to handle the game loop. For specific operations,
it must use some specific classes/namespaces to do that properly.

--