audio handler: fix panning effect on windows
Original problem:

```
Panninganning effects are only available on the Linux version because it works properly.
On Windows, I have a random crash and I was not able to find the bug.

It is possible the bug comes from LunatiX, or SDL2 (SDL 2.0.3), but I am not sure.
```

Update:

Upgrading SDL2 mixer to 2.0.5 fixed the bug. Now you can have panning effect on Windows.

References:
- commit 795f226

--