game controller database: update
The new beta version of LunatiX 0.14.0 can read the latest version of the database (gamecontrollerdb.txt).
This version of the database is used if LunatiX was compiled with SDL 2.0.6+.

Since LunatiX LunatiX requires at least SDL 2.0.5,
the database related to this specific version of SDL annot work on SDL 2.0.6+.

That is why I added gamecontrollerdb_205.txt in the configuration file,
so LunatiX can retrieve the good file.

More details:

- [Game controller Database](https://github.com/gabomdq/SDL_GameControllerDB)
- [LunatiX/https://github.com/Gumichan01/lunatix/commit/2f73432](https://github.com/Gumichan01/lunatix/commit/2f73432942061eac5cf3638076022e74911df91a)

--