i18n: user-defined language
Using the system environment language is a good idea on Linux.
But it does not work on Windows (unless you launch the game from Git Bash).
That is why I use the language defined in the XML file (asset.xml).

By default, the language defined by the user in the XML file is chosen over
the language set in the environnement system.

→ issue #21